cheroot/requirements/tox-linkcheck-docs-cp38-linux-x86_64.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --allow-unsafe --output-file=requirements/tox-linkcheck-docs-cp38-linux-x86_64.txt --strip-extras requirements/tox-linkcheck-docs-cp38-linux-x86_64.in setup.cfg
#
alabaster==0.7.12
    # via sphinx
attrs==21.4.0
    # via pytest
babel==2.9.1
    # via sphinx
beautifulsoup4==4.10.0
    # via furo
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via -r requirements/tests.in
charset-normalizer==2.0.10
    # via requests
colorama==0.4.4
    # via pytest-watch
coverage==6.2 ; python_version >= "3.6"
    # via
    #   -r requirements/tests.in
    #   pytest-cov
cryptography==36.0.1
    # via
    #   pyopenssl
    #   trustme
docopt==0.6.2
    # via pytest-watch
docutils==0.16
    # via
    #   sphinx
    #   sphinx-tabs
execnet==1.9.0
    # via pytest-xdist
furo==2022.1.2
    # via -r requirements/tox-docs.in
idna==3.3
    # via
    #   requests
    #   trustme
imagesize==1.3.0
    # via sphinx
importlib-resources==5.4.0
    # via jaraco.text
iniconfig==1.1.1
    # via pytest
jaraco.context==4.1.1
    # via -r requirements/tests.in
jaraco.functools==3.5.0
    # via
    #   cheroot (setup.cfg)
    #   jaraco.text
    #   tempora
jaraco.packaging==8.2.1
    # via -r requirements/tox-docs.in
jaraco.text==3.6.0
    # via -r requirements/tests.in
jinja2==3.0.3
    # via sphinx
markupsafe==2.0.1
    # via jinja2
more-itertools==8.12.0 ; python_version >= "3.6"
    # via
    #   cheroot (setup.cfg)
    #   jaraco.functools
packaging==21.3
    # via
    #   pytest
    #   pytest-sugar
    #   sphinx
pbr==5.8.0
    # via sphinxcontrib-apidoc
pluggy==1.0.0
    # via pytest
portend==3.1.0
    # via -r requirements/tests.in
py==1.11.0
    # via
    #   pypytools
    #   pytest
    #   pytest-forked
pycparser==2.21
    # via cffi
pyenchant==3.2.2
    # via sphinxcontrib-spelling
pygments==2.11.2
    # via
    #   furo
    #   sphinx
    #   sphinx-tabs
pyopenssl==21.0.0
    # via -r requirements/tests.in
pyparsing==3.0.7
    # via packaging
pypytools==0.6.2
    # via -r requirements/tests.in
pytest==6.2.5
    # via
    #   -r requirements/tests.in
    #   pytest-cov
    #   pytest-forked
    #   pytest-mock
    #   pytest-rerunfailures
    #   pytest-sugar
    #   pytest-watch
    #   pytest-xdist
pytest-cov==2.12.0 ; python_version != "3.4"
    # via -r requirements/tests.in
pytest-forked==1.4.0 ; sys_platform != "win32" and (python_version < "3.0" or python_version > "3.4")
    # via
    #   -r requirements/tests.in
    #   pytest-xdist
pytest-mock==3.6.1
    # via -r requirements/tests.in
pytest-rerunfailures==10.2 ; python_version >= "3.6"
    # via -r requirements/tests.in
pytest-sugar==0.9.4
    # via -r requirements/tests.in
pytest-watch==4.2.0
    # via -r requirements/tests.in
pytest-xdist==2.5.0
    # via -r requirements/tests.in
python-dateutil==2.8.2
    # via -r requirements/tox-docs.in
pytz==2021.3
    # via
    #   babel
    #   tempora
requests==2.27.1
    # via
    #   requests-toolbelt
    #   requests-unixsocket
    #   sphinx
requests-toolbelt==0.9.1
    # via -r requirements/tests.in
requests-unixsocket==0.3.0
    # via -r requirements/tests.in
six==1.16.0
    # via
    #   cheroot (setup.cfg)
    #   pyopenssl
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.3.1
    # via beautifulsoup4
sphinx==4.3.2
    # via
    #   -r requirements/tox-docs.in
    #   furo
    #   sphinx-tabs
    #   sphinxcontrib-apidoc
    #   sphinxcontrib-spelling
sphinx-tabs==3.2.0
    # via -r requirements/tox-docs.in
sphinxcontrib-apidoc==0.3.0
    # via -r requirements/tox-docs.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sphinxcontrib-spelling==7.3.2
    # via -r requirements/tox-linkcheck-docs-cp38-linux-x86_64.in
tempora==5.0.0
    # via portend
termcolor==1.1.0
    # via pytest-sugar
toml==0.10.2
    # via pytest
tomli==2.0.0
    # via coverage
trustme==0.9.0
    # via -r requirements/tests.in
urllib3==1.26.8 ; python_version != "3.4"
    # via
    #   -r requirements/tests.in
    #   requests
watchdog==2.1.6
    # via pytest-watch
zipp==3.7.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
setuptools==60.5.0
    # via
    #   jaraco.packaging
    #   pytest-rerunfailures
    #   sphinx
